@@ -1,5 +1,6 @@
 import React from 'react';
 import { Remarkable } from 'remarkable';
+import { HOURS_FORMAT } from '../../common/constants';
 
 const markdown = (text) => {
   const md = new Remarkable();
@@ -21,13 +22,13 @@ const Event = ({ title, start_time, end_time, content, eventClickHandler, index,
         }}
         className="cal-event-header"
       >
-        <p className="cal-event-date">{Date.parse(start_time).toLocaleString('no-nb')}</p>
+        <p className="cal-event-date">{HOURS_FORMAT.format(new Date(start_time))}</p>
         <h3 className="cal-event-title">{title}</h3>
       </div>
 
       <div className={`cal-event-content ${classes}`} dangerouslySetInnerHTML={markdown(content)} />
       <p className={`cal-event-content ${classes} cal-event-endtime`}>
-        Antatt sluttidspunkt: {Date.parse(end_time).toLocaleString('no-nb')}
+        Antatt sluttidspunkt: {HOURS_FORMAT.format(new Date(end_time))}
       </p>
     </div>
   );
